+++ src/App.js
@@ -2,18 +2,17 @@ import React from 'react';
 import { BrowserRouter } from 'react-router-dom';
 import './App.css';
 import 'materialize-css';
-import Authenticate from './components/context/Authenticate'
+import Authenticate from './components/context/Authenticate';
 
-import Router from './components/routes/Router'
+import Router from './components/routes/Router';
 
 function App() {
 	return (
 		<div className="App">
 			<BrowserRouter>
-			<Authenticate>
-			<Router />
-			</Authenticate>
-				
+				<Authenticate>
+					<Router />
+				</Authenticate>
 			</BrowserRouter>
 		</div>
 	);
